+++ res/libsodium/buildlibsodium.sh
@@ -1,14 +1,13 @@
 #!/bin/bash
 cd "$(dirname "$0")/../.."
 set -e
+SCRIPT_DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")/../.." && pwd)/src/scripts"
+ME="buildlibsodium"
+. "$SCRIPT_DIR/lib-log.sh"
 
-# First thing to do is see if libsodium.a exists in the res folder. If it does, then there's nothing to do
-if [ -f res/libsodium.a ]; then
-    rm res/libsodium.a
-    rm -rf res/libsodium/libsodium-1.0.21
-fi
+[ -f res/libsodium.a ] && { rm res/libsodium.a; rm -rf res/libsodium/libsodium-1.0.21; }
 
-echo "Building libsodium"
+notice "Building libsodium..."
 
 # Go into the lib sodium directory
 cd res/libsodium
@@ -37,6 +36,5 @@ if [ -f libsodium-1.0.21/src/libsodium/.libs/libsodium.a ]; then
 elif [ -f libsodium-1.0.21/.libs/libsodium.a ]; then
     cp libsodium-1.0.21/.libs/libsodium.a ../
 else
-    echo "ERROR: libsodium.a not found after build"
-    exit 1
+    err "libsodium.a not found after build"
 fi
